Remove old enter alias, add new enter function
This allows the "enter" command to always get the latest docker image.
Previously, the enter alias was evaluated on login and if the latest docker
container changed, you either had to logout/login or source ~/.bashrc

Now it is always evaluated when the command is ran, giving you the latest
container all the time.