Run.sh
A generic program to run the specified command in the Docker machine
as the Anthem user. Useful for shell scripting. It's designed so that
Chris's station can ssh into the display, run the get_sn_all.py
script, and exit without any human interaction. The enter alias
gets into the Docker machine, but doesn't allow a command to be executed.